Implement function for getting route's scope info from server/config.
